backend/requirements.txt
fastapi 
uvicorn 
langchain-text-splitters 
langchain-chroma 
langchain-community 
google-generativeai 
pypdf 
python-dotenv
